added and tested remove single Post REST call

+++ app.py
@@ -111,10 +111,26 @@ def createPost():
 @app.route('/singlepost', methods=['DELETE'])
 def deletePost():
     db = manager.getDBConnection()
-    body = request.json
-    # do something, eg. return json response
-    return postAPI.delete_post(db, body)
+    body = request.args
+    if not body:
+        return Response("{ 'Result': 'Error: No arguments given' }", status=400, mimetype='application/json')
+    correctArgs = ['stock', 'postID']
+    validated = True
+    for key in body:
+        if key not in correctArgs:
+            validated = False
+            break
 
+    if not validated or len(body) != 2:
+        return Response("{ 'Result': 'Error: Bad args given' }", status=400, mimetype='application/json')
+    # do something, eg. return json response
+    result = postAPI.delete_post(db, body)
+    if result == 0:
+        return Response("{ 'Result': 'Removed Post' }", status=200, mimetype='application/json')
+    elif result == 1:
+        return Response("{ 'Result': 'Error: Could Not Find Tagged Post with given ID' }", status=400, mimetype='application/json')
+    else:
+        return Response("{ 'Result': 'Unknown Error' }", status=500, mimetype='application/json')
 
 if __name__ == '__main__':
     app.run(debug=True)
